// src/store/products/products.enum.ts
enum ProductsActionTypeEnum {
  GET_PRODUCTS_PENDING = "GET_PRODUCTS_PENDING",
  GET_PRODUCTS_FAILED = "GET_PRODUCTS_FAILED",
  GET_PRODUCTS_SUCCESS = "GET_PRODUCTS_SUCCESS",

  GET_PRODUCT_BY_ID_SUCCESS = "GET_PRODUCT_BY_ID_SUCCESS",
  GET_PRODUCT_BY_ID_PENDING = "GET_PRODUCT_BY_ID_PENDING",
  GET_PRODUCT_BY_ID_FAILED = "GET_PRODUCT_BY_ID_FAILED",

  ADD_PRODUCTS_PENDING = "ADD_PRODUCTS_PENDING",
  ADD_PRODUCTS_FAILED = "ADD_PRODUCTS_FAILED",
  ADD_PRODUCTS_SUCCESS = "ADD_PRODUCTS_SUCCESS",

  UPDATE_PRODUCTS_PENDING = "UPDATE_PRODUCTS_PENDING",
  UPDATE_PRODUCTS_FAILED = "UPDATE_PRODUCTS_FAILED",
  UPDATE_PRODUCTS_SUCCESS = "UPDATE_PRODUCTS_SUCCESS",

  DELETE_PRODUCTS_PENDING = "DELETE_PRODUCTS_PENDING",
  DELETE_PRODUCTS_FAILED = "DELETE_PRODUCTS_FAILED",
  DELETE_PRODUCTS_SUCCESS = "DELETE_PRODUCTS_SUCCESS",
}

export default ProductsActionTypeEnum;